requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
beautifulsoup4==4.6.3
bleach==2.1.3
bs4==0.0.1
cffi==1.11.5
click==6.7
cryptography==2.3.1
dominate==2.3.1
enum34==1.1.6
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask_Login==0.4.1
Flask_Mail==0.9.1
Flask_Moment==0.6.0
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask_WTF==0.14.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
micawber==0.3.5
peewee==3.5.1
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
python-http-client==3.1.0
pyjwt==1.5.3
sendgrid==5.6.0
six==1.11.0
SQLAlchemy==1.2.11
virtualenv==16.0.0
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.14.1
